challenge1b_tinyllama_full/requirements.txt
pytesseract
PyMuPDF
Pillow
sentence-transformers
nltk

transformers
accelerate
sentencepiece